Update #5_For Loop in c++ Advance || Prime Number || Factorial || Fibonacci || Sum of n number
Day 5/180: For Loop in advance
Problems: 

Print number from 280 to 250 with the help of for loop.
Print char from ‘A’ to ‘Z’ with the help of a for loop.
Print char from ‘Z’ to ‘A’ with the help of a for loop.
There is an Arithmetic Progression, First number is 220 and Common Difference is 7. So print all the numbers from 220 to 730 which follow AP.
Print Sum of square of first n natural number.
Print Sum of cube of first n natural number
Print n’th Fibonacci number.